+++ lib/widgets/tools/images_and_files/hexstring2file.dart
@@ -8,6 +8,7 @@ import 'package:gc_wizard/i18n/app_localizations.dart';
 import 'package:gc_wizard/logic/tools/images_and_files/hexstring2file.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_button.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_iconbutton.dart';
+import 'package:gc_wizard/widgets/common/base/gcw_output_text.dart';
 import 'package:gc_wizard/widgets/common/base/gcw_textfield.dart';
 import 'package:gc_wizard/widgets/common/gcw_default_output.dart';
 import 'package:gc_wizard/widgets/common/gcw_exported_file_dialog.dart';
@@ -91,15 +92,13 @@ class HexString2FileState extends State<HexString2File> {
 
       if (_outData == null) return null;
 
-      return hexDataOutput(context, _outData);
+      return hexDataOutput(context, <Uint8List>[_outData]);
     } else {
       if (_outData == null) return null;
 
       return file2hexstring(_outData);
     }
   }
-
-
 
   _exportFile(BuildContext context, Uint8List data) async {
     var fileType = getFileType(data);
@@ -110,73 +109,62 @@ class HexString2FileState extends State<HexString2File> {
   }
 }
 
-hexDataOutput(BuildContext context, Uint8List _outData) {
-  if (_outData == null) return '';
-  var mimeType = getMimeType(getFileType(_outData));
+Widget hexDataOutput(BuildContext context, List<Uint8List> outData) {
+  if (outData == null) return Container();
 
-  switch (mimeType) {
-    case MIMETYPE.IMAGE:
-      try {
-        return Image.memory(_outData);
-      } catch (e) {
-        return getFileType(_outData).replaceFirst('.', '') + '-' +
-            i18n(context, 'hexstring2file_file');
-      }
+  var children = outData.map((_outData) {
+    var mimeType = getMimeType(getFileType(_outData));
 
-      return null;
-
-    case MIMETYPE.TEXT:
-      return String.fromCharCodes(_outData);
-
-    case MIMETYPE.ARCHIV:
-      String fileNames = '';
-      String extension = getFileType(_outData);
-      if (extension.endsWith('.zip')) {
+    switch (mimeType) {
+      case MIMETYPE.IMAGE:
         try {
-          InputStream input = new InputStream(_outData.buffer.asByteData());
-          // Decode the Zip file
-          final archive = ZipDecoder().decodeBuffer(input);
-          fileNames = archive.map((file) {
-            if (file.isFile)
-              return ('-> ' + file.name);
-            else
-              return '';
-          }).join('\n');
-        } catch (e) {}
+          return Image.memory(_outData);
+        } catch (e) {
+        }
+        return _fileWidget(context, getFileType(_outData));
 
-        return 'zip-' +
-            i18n(context, 'hexstring2file_file') +
-            ' -> ' +
-            i18n(context, 'hexstring2file_content') +
-            '\n' +
-            fileNames;
-      } else if (extension.endsWith('.tar')) {
-        try {
-          InputStream input = new InputStream(_outData.buffer.asByteData());
-          // Decode the Zip file
-          final archive = TarDecoder().decodeBuffer(input);
-          fileNames = archive.map((file) {
-            if (file.isFile)
-              return ('-> ' + file.name);
-            else
-              return '';
-          }).join('\n');
-        } catch (e) {}
+      case MIMETYPE.TEXT:
+        return GCWOutputText(text: String.fromCharCodes(_outData));
 
-        return 'tar-' +
-            i18n(context, 'hexstring2file_file') +
-            ' -> ' +
-            i18n(context, 'hexstring2file_content') +
-            '\n' +
-            fileNames;
-      } else {
-        fileNames = extension.replaceFirst('.', '') + '-' +
-            i18n(context, 'hexstring2file_file');
-      }
+      case MIMETYPE.ARCHIV:
+        String extension = getFileType(_outData);
+        if (extension.endsWith('.zip')) {
+          try {
+            InputStream input = new InputStream(_outData.buffer.asByteData());
+            return (_archiveWidget(context, ZipDecoder().decodeBuffer(input) , 'zip'));
+          } catch (e) {}
+        } else if (extension.endsWith('.tar')) {
+          try {
+            InputStream input = new InputStream(_outData.buffer.asByteData());
+            return (_archiveWidget(context, TarDecoder().decodeBuffer(input), 'tar'));
+          } catch (e) {}
 
-      return fileNames;
-    default:
-      return getFileType(_outData).replaceFirst('.', '') + '-' +
-          i18n(context, 'hexstring2file_file');
-  }
+        } else {
+          return _fileWidget(context, extension);
+        }
+        break;
+      default:
+        return _fileWidget(context, getFileType(_outData));
+    }
+    return Container();
+  }).toList();
+
+  return Column(children: children);
 }
+
+Widget _archiveWidget(BuildContext context, Archive archive, String type) {
+  var text  = type + "-" + i18n(context, 'hexstring2file_file') + ' -> ' +
+      i18n(context, 'hexstring2file_content') + '\n';
+
+  text += archive.where((element) => element.isFile).map((file) {
+    return ('-> ' + file.name);
+  }).join('\n');
+
+  return GCWOutputText(text: text);
+}
+
+Widget _fileWidget(BuildContext context, String fileName) {
+  return GCWOutputText(text: fileName.replaceFirst('.', '') + '-' +
+        i18n(context, 'hexstring2file_file'));
+}
+
